Modified test.sh to exclude tests in test/draw/gpu when using the nogpu argument.

+++ test.sh
@@ -8,6 +8,8 @@ fi
 FLAGS=${@:2}
 TESTS=$@
 case "$1" in
+	"nogpu")
+		;&
 	"")
 		;&
 	"all")
@@ -25,22 +27,25 @@ else
 fi
 for TEST in $TESTS
 do
-	OUTPUT=$(rock -q -lpthread --gc=off $ARGS $FLAGS $TEST 2>&1 )
-	if [[ !( $? == 0 ) ]]
+	if [[ !( $1 == "nogpu" && $TEST == *"/gpu/"* ) ]]
 	then
-		echo -e $OUTPUT
-		echo "Compilation failed of test: $TEST"
-		exit 2
+		OUTPUT=$(rock -q -lpthread --gc=off $ARGS $FLAGS $TEST 2>&1 )
+		if [[ !( $? == 0 ) ]]
+		then
+			echo -e $OUTPUT
+			echo "Compilation failed of test: $TEST"
+			exit 2
+		fi
+		NAME=${TEST##*/}
+		NAME=${NAME%.*}
+		#echo $TEST
+		./$NAME
+		if [[ !( $? == 0 ) ]]
+		then
+			echo "Failed test: $TEST"
+			exit 1
+		fi
+		rm -f $NAME
 	fi
-	NAME=${TEST##*/}
-	NAME=${NAME%.*}
-	#echo $TEST
-	./$NAME
-	if [[ !( $? == 0 ) ]]
-	then
-		echo "Failed test: $TEST"
-		exit 1
-	fi
-	rm -f $NAME
 done
 exit 0
